+++ kandy_sample_reactnative_ios/ios/kandy_sample_reactnative_ios/Persence.m
